+++ src/main/g8/backend/src/main/scala/$organization;format="packaged"$/Main.scala
@@ -1,7 +1,11 @@
 package $organization$
 
-object Main extends App {
+object Hello {
   val hello = "Hello World!"
+}
+
+object Main extends App {
+  val hello = Hello.hello
   println(hello)
 }