+++ src/main/g8/backend/src/test/scala/$organization;format="packaged"$/MainSpec.scala
@@ -4,7 +4,7 @@ import org.scalatest.FlatSpec
 import org.scalatest.Matchers
 
 class MainSpec extends FlatSpec with Matchers {
-  "the main class" should "have a hello message" in {
-    Main.hello should be "Hello World!"
+  "the hello object" should "have a hello message" in {
+    Hello.hello should be("Hello World!")
   }
 }