+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### [1.0.0]
 #### Added
 - Removed play framework and slick, added doobie and finch
+- Added unit tests for finch and backend
 
 ### [0.0.2]
 #### Fixed